requirements.txt
Flask==2.2.5
python-dotenv==0.18.0
requests==2.32.4
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-grpc
